requirements.txt
flask
llama-cpp-python
sentence-transformers
chromadb
